Add systemless wlan.ko injection
Works with both stock OxygenOS and Oreo 8.1 custom roms
regardless of being rooted or not. We inject the proper
SELinux policies during Kernel flash so the bind mount will work.

Extra work was needed for Oreo 8.1 because /vendor is mounted up early
in the boot process which is why there are sepolicies for modprobe
and the extra modprobe call after the mount bind. On OxygenOS
these are ignored.

Thanks to osm0sis once more for figuring pretty much how to achieve
this without root and cleanly.

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>